refactor(translations): update download script to esm
