+++ docs/docs/contributing/code/reference/i18n.md
@@ -31,7 +31,7 @@ Periodically, we export a snapshot of the translated strings from Transifex and 
 You can update locale files by running this on the command line:
 
 ```shell-session
-node ./bin/download_translations.js
+node ./bin/download_translations.mjs
 ```
 
 :::info
